backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0

# Database
pymongo==4.6.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Rate Limiting
Flask-Limiter==3.5.0

# Environment Variables
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
